Move IO code to examples directory
