+++ sdks/go/examples/native_wordcap/nativepubsubio/subscriptiontracker.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package pubsubio
+package nativepubsubio
 
 import "math"
