+++ sdks/go/pkg/beam/io/pubsubio/pubsubio.go
@@ -26,7 +26,6 @@ import (
 	"reflect"
 
 	"github.com/apache/beam/sdks/v2/go/pkg/beam"
-	"github.com/apache/beam/sdks/v2/go/pkg/beam/core/sdf"
 	"github.com/apache/beam/sdks/v2/go/pkg/beam/core/typex"
 	"github.com/apache/beam/sdks/v2/go/pkg/beam/core/util/protox"
 	"github.com/apache/beam/sdks/v2/go/pkg/beam/core/util/reflectx"
@@ -47,10 +46,7 @@ func init() {
 	register.Function2x1(marshalMessageFn)
 	register.Function2x0(wrapInMessage)
 	register.Function2x0(wrapInMessage)
-	register.DoFn4x2[beam.BundleFinalization, *sdf.LockRTracker, []byte, func(beam.EventTime, []byte), sdf.ProcessContinuation, error](&PubSubRead{})
-	register.DoFn1x1[[]byte, error](&PubSubWrite{})
 	register.Emitter1[[]byte]()
-	register.Emitter2[beam.EventTime, []byte]()
 	register.Emitter1[*pb.PubsubMessage]()
 }
